+++ packages/bento-design-system/src/Modal/Config.ts
@@ -1,7 +1,7 @@
 import { IconProps } from "../Icons";
 import { BentoSprinkles } from "../internal";
 import { TitleProps } from "../Typography/Title/Title";
-import { ModalKind, ModalSize } from "./createModal";
+import { ModalKind, ModalSize } from "./Modal";
 import { Children } from "../util/Children";
 import { ActionsProps } from "../Actions/Actions";
